Include new c10d frontend in build
