+++ torch/csrc/distributed/c10d/comm.h
@@ -4,12 +4,12 @@
 
 #include <ATen/ATen.h>
 #include <c10d/ProcessGroup.hpp>
-#include <torch/csrc/utils/pybind.h>
+#include <ATen/core/ivalue_inl.h>
 
 namespace c10d {
 
 // Broadcast many tensors to all processes in the process group.
-void broadcast_coalesced(
+TORCH_API void broadcast_coalesced(
     std::shared_ptr<c10d::ProcessGroup> process_group,
     at::TensorList tensors,
     size_t buffer_size,
@@ -47,7 +47,7 @@ struct TORCH_PYTHON_API CommHookInterface {
   // the bucket is ready. The hook can perform whatever processing is
   // needed and return a Future that will hold the new value of the grad
   // bucket's tensors once ready.
-  virtual c10::intrusive_ptr<torch::jit::Future> runHook(
+  virtual c10::intrusive_ptr<c10::ivalue::Future> runHook(
       const GradBucket& bucket) = 0;
 
   // Once the grad bucket of Future is ready, c10d reducer will call this
@@ -57,36 +57,5 @@ struct TORCH_PYTHON_API CommHookInterface {
   virtual std::vector<at::Tensor> processFuture(c10::IValue future_value) = 0;
 };
 
-// PythonCommHook enables registering a python hook to c10d reducer and is a
-// sub class of CommHookInterface.
-class TORCH_PYTHON_API PythonCommHook : public CommHookInterface {
- public:
-  // The constructor takes a state and a callable hook. Inputs are Python
-  // objects. The state is passed to the hook in runHook function can be used to
-  // maintain and update any state information that users would like to maintain
-  // as part of the training process. The hook can perform whatever processing
-  // user specified and return a Future indicating completion of any async work.
-  PythonCommHook(py::object state, py::object hook);
-
-  ~PythonCommHook() override {
-    py::gil_scoped_acquire ag;
-    state_.dec_ref();
-    hook_.dec_ref();
-    // explicitly setting PyObject* state_ and hook_ to nullptr to prevent
-    // py::object's dtor to decref on the PyObject again.
-    // See Note [Destructing py::object] in python_ivalue.h
-    state_.ptr() = nullptr;
-    hook_.ptr() = nullptr;
-  }
-
-  c10::intrusive_ptr<torch::jit::Future> runHook(
-      const GradBucket& bucket) override;
-
-  std::vector<at::Tensor> processFuture(c10::IValue future_value) override;
-
- private:
-  py::object state_;
-  py::object hook_;
-};
 
 } // namespace c10d
